+++ modules/contracts/src.ts/tasks/index.ts
@@ -1,4 +1,5 @@
 import "./createChannel";
 import "./displayAccounts";
+import "./drip";
 import "./registerTransfer";
 import "./changeTransferRegistryOwner";